requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
customtkinter==5.2.2
darkdetect==0.8.0
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
onnxruntime==1.19.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
ppocr-onnx==0.0.3.9
protobuf==5.28.2
psutil==6.1.0
pyasn1==0.6.1
pyclipper==1.3.0.post5
pypiwin32==223
pyreadline3==3.5.4
pywin32==306
requests==2.32.3
rsa==4.9
setuptools==75.6.0
shapely==2.0.6
sympy==1.13.3
urllib3==2.2.3
win11toast==0.35
winsdk==1.0.0b10
